[Code]: more clean up to the test option
