+++ x-pack/plugins/code/server/__tests__/lsp_indexer_test.ts
@@ -19,8 +19,7 @@ import { Logger } from '../log';
 import { InstallManager } from '../lsp/install_manager';
 import { LspService } from '../lsp/lsp_service';
 import { RepositoryConfigController } from '../repository_config_controller';
-import { ServerOptions } from '../server_options';
-import { emptyAsyncFunc, TEST_OPTIONS } from '../test_utils';
+import { createTestServerOption, emptyAsyncFunc } from '../test_utils';
 import { ConsoleLoggerFactory } from '../utils/console_logger_factory';
 
 const log: Logger = new ConsoleLoggerFactory().getLogger(['test']);
@@ -60,15 +59,7 @@ function prepareProject(url: string, p: string) {
 
 const repoUri = 'github.com/Microsoft/TypeScript-Node-Starter';
 
-const config = {
-  get(key: string) {
-    if (key === 'path.data') {
-      return '/tmp/test';
-    }
-  },
-};
-
-const serverOptions = new ServerOptions(TEST_OPTIONS, config);
+const serverOptions = createTestServerOption();
 
 function cleanWorkspace() {
   return new Promise(resolve => {
